Add TmpPortal component to root route
This commit introduces a new TmpPortal component and sets it as the landing component of the application by adding it to the root ("/") route in App.js. This temporary portal includes a button routing to the routine configuration page. The changes remove usage of RoutineDaysConfigurator and TrainingForm from App.js as they are not needed for the purpose.